LPS-60850 Move to internal package
