+++ modules/apps/collaboration/wiki/wiki-service/src/main/java/com/liferay/wiki/internal/convert/creole/WikiCreoleConvertProcess.java
@@ -12,16 +12,16 @@
  * details.
  */
 
-package com.liferay.wiki.convert.creole;
+package com.liferay.wiki.internal.convert.creole;
 
 import com.liferay.portal.convert.BaseConvertProcess;
 import com.liferay.portal.convert.ConvertProcess;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.util.MaintenanceUtil;
+import com.liferay.wiki.internal.translator.ClassicToCreoleTranslator;
 import com.liferay.wiki.model.WikiPage;
 import com.liferay.wiki.service.WikiPageLocalService;
-import com.liferay.wiki.translator.ClassicToCreoleTranslator;
 
 import java.util.List;
