+++ modules/apps/collaboration/wiki/wiki-service/src/main/java/com/liferay/wiki/internal/translator/ClassicToCreoleTranslator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.translator;
+package com.liferay.wiki.internal.translator;
 
 /**
  * @author Jorge Ferrer
