+++ modules/apps/collaboration/wiki/wiki-service/src/main/java/com/liferay/wiki/internal/translator/MediaWikiToCreoleTranslator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.wiki.translator;
+package com.liferay.wiki.internal.translator;
 
 import com.liferay.portal.kernel.util.CharPool;
 import com.liferay.portal.kernel.util.StringPool;
